refactor(gpu_gkr): clippy-clean (0 own-crate warnings)
Biggest crate in the split (150 own-crate warnings, matching the
deep-clean baseline note that most post-split warnings landed here).

Structural / crate-level fixes:
- Fixed a deny-level clippy error (`not_unsafe_ptr_arg_deref` on
  `schedule_execute_main_layer_from_workflow_state`'s device-pointer
  params) with an allow + opaque-FFI-handle justification, matching
  gpu_core's `compute_minimal_carveout` precedent.
- Crate-level `#![allow(clippy::too_many_arguments)]` for the kernel
  launcher-shaped scheduling functions (same precedent as gpu_hash /
  gpu_ntt / gpu_execution_prover / gpu_trace).
- Crate-level `#![allow(private_bounds)]`: `ForwardKernels` /
  `BackwardKernels` / `SetupKernels` are deliberately sealed per-phase
  kernel-dispatch traits (pub(crate) by design per lib.rs's own
  `GpuKernels` comment); every private_bounds warning in the crate
  traced back to these three. Escalation item (see task-13 report).

private_interfaces resolutions (grepped gpu_whir/gpu_circuit_prover for
external consumers before choosing):
- `ProofLayout`'s `output_evaluations`/`backward`/`whir` fields demoted
  to `pub(crate)` (matching their already-`pub(crate)` element types) —
  every cross-crate consumer goes through the `impl ProofLayout`
  accessors, never the raw fields.
- `GpuBaseFieldSourceKind` widened to `#[doc(hidden)] pub` — it's the
  `source_kind` field type on two `#[doc(hidden)] pub` FFI descriptors
  apex tests D2H-copy and inspect field-by-field; narrower would leave
  one field un-inspectable on an otherwise uniform descriptor.
- `ImmediateFactorRecipeStructural::monomials()` demoted to
  `pub(crate)` — nothing outside the crate calls it; `key()` (plain
  tuples) is the sanctioned public introspection API.

Other categories: unused imports (some `#[cfg(test)]`-gated per the
established upstream.rs convention rather than deleted), unnecessary
casts, field_reassign_with_default, needless_range_loop (verified
partial-range semantics per case, one legitimate false-positive avoided
in setup/tests.rs), needless_borrow/pass_by_value, derivable_impls,
len_without_is_empty, type_complexity (factored 3 type aliases),
map_entry, drop_non_drop (removed 2 genuinely-inert `drop()` calls,
confirmed via cargo check that NLL still ends the borrows), and a
handful of one-off doc/naming lints.

e2e parity pair (run_add_sub_proof_parity_test{,_sec100}) re-run and
PASS: production scheduling files were touched throughout (forward/mod.rs,
backward/mod.rs, stage1/mod.rs, storage/ops.rs, sumcheck_plan.rs, …).

+++ gpu/gkr/src/backward/builders.rs
@@ -195,7 +195,7 @@ fn memory_query_as_flattened_relation<E: Field + FieldExtension<BF>>(
                 let offset_challenge = ImmediateCoeff::challenge_scaled(
                     idx as u8,
                     external_challenges.permutation_argument_linearization_challenges[idx],
-                    BF::from_u32_unchecked(*low_offset as u32),
+                    BF::from_u32_unchecked(*low_offset),
                 );
                 constant_term.add_assign(&offset_challenge);
             }
@@ -227,7 +227,7 @@ fn memory_query_as_flattened_relation<E: Field + FieldExtension<BF>>(
                 let offset_challenge = ImmediateCoeff::challenge_scaled(
                     idx as u8,
                     external_challenges.permutation_argument_linearization_challenges[idx],
-                    BF::from_u32_unchecked(rel.timestamp_offset as u32),
+                    BF::from_u32_unchecked(rel.timestamp_offset),
                 );
                 constant_term.add_assign(&offset_challenge);
             }
